1) Update numpy to 1.24.1 and address a deprecation of numpy type aliases. 2) Update GitHub Actions + README to Python 3.8 (not setup.py yet, since it might be best to not mess with that till this round of updates is finished).

+++ spacer/caffe_utils.py
@@ -83,7 +83,7 @@ def classify_from_imlist(im_list: List,
                         transformer.preprocess(im_list[pos])
             net.forward(start=startlayer)
             scorelist.extend(list(copy(net.blobs[scorelayer].data).
-                                  astype(np.float)))
+                                  astype(float)))
 
         scorelist = scorelist[:len(im_list)]
